requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.1
numpy==2.2.3
matplotlib==3.10.1
pillow==11.1.0
requests==2.32.3
opencv-python==4.9.0.80
insightface==0.7.3
onnxruntime==1.15.1
onnx==1.14.1
scikit-learn==1.3.2
openai==1.12.0 